refactor(scroll): 스크롤 애니메이션 공통 함수 추출 및 통일
