+++ app/_components/ui/ScrollToTop.tsx
@@ -2,6 +2,7 @@
 
 import { useState, useEffect, useCallback } from 'react';
 import { FiArrowUp } from 'react-icons/fi';
+import { smoothScrollToTop } from '@/_lib/utils/scroll';
 
 interface ScrollToTopProps {
     containerRef?: React.RefObject<HTMLElement | null>;
@@ -39,38 +40,7 @@ export default function ScrollToTop({ containerRef, threshold = 400 }: ScrollToT
     }, [containerRef, toggleVisibility]);
 
     const scrollToTop = () => {
-        const isWindow = !containerRef?.current;
-        const startPosition = isWindow ? window.scrollY : containerRef!.current!.scrollTop;
-
-        if (startPosition === 0) return;
-
-        const duration = 2000; // 2000ms = 2초
-        let startTime: number | null = null;
-
-        // Easing function: easeOutCubic
-        // 1 - (1 - t)^3
-        const easeOutCubic = (t: number) => 1 - Math.pow(1 - t, 3);
-
-        const animation = (currentTime: number) => {
-            if (startTime === null) startTime = currentTime;
-            const timeElapsed = currentTime - startTime;
-            const progress = Math.min(timeElapsed / duration, 1);
-            const ease = easeOutCubic(progress);
-
-            const nextScroll = startPosition * (1 - ease);
-
-            if (isWindow) {
-                window.scrollTo(0, nextScroll);
-            } else {
-                containerRef!.current!.scrollTop = nextScroll;
-            }
-
-            if (timeElapsed < duration) {
-                requestAnimationFrame(animation);
-            }
-        };
-
-        requestAnimationFrame(animation);
+        smoothScrollToTop(containerRef?.current ?? null);
     };
 
     return (
